feat: created all CRUD operations for Layer

+++ src/main/java/vlad/erofeev/layerservice/repositories/LayerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import vlad.erofeev.layerservice.domain.entities.Layer;
 
 @Repository
-public interface LayerRepository extends JpaRepository<Layer, Integer> {
+public interface LayerRepository extends JpaRepository<Layer, Long> {
 }
